requirements.txt
streamlit
opencv-python
python-docx
numpy
scikit-learn
joblib==1.4.2
gdown==4.4.0
lxml
xmltodict
opencv-python-headless
matplotlib
PyPDF2
numpy
pillow
